chore(backup): remove waku backup and keep only local backup
Needed for https://github.com/status-im/status-desktop/issues/18123
