+++ api/defaults.go
@@ -48,7 +48,6 @@ func defaultSettings(keyUID string, address string, derivedAddresses map[string]
 	chatKeyString := derivedAddresses[accscommon.PathEIP1581Chat].PublicKey
 
 	s := &settings.Settings{}
-	s.BackupEnabled = true
 	logLevel := "INFO"
 	s.LogLevel = &logLevel
 	s.ProfilePicturesShowTo = settings.ProfilePicturesShowToEveryone
@@ -294,7 +293,6 @@ func DefaultNodeConfig(installationID, keyUID string, request *requests.CreateAc
 	nodeConfig.LogDir = request.LogFilePath
 	nodeConfig.LogLevel = DefaultLogLevel
 	nodeConfig.DataDir = DefaultDataDir
-	nodeConfig.ProcessBackedupMessages = false
 	nodeConfig.KeycardPairingDataFile = DefaultKeycardPairingDataFile
 	if request.KeycardPairingDataFile != nil {
 		nodeConfig.KeycardPairingDataFile = *request.KeycardPairingDataFile
