+++ multiaccounts/settings/database.go
@@ -149,11 +149,10 @@ INSERT INTO settings (
   test_networks_enabled,
   fleet,
   auto_refresh_tokens_enabled,
-  news_feed_last_fetched_timestamp,
-  backup_fetched
+  news_feed_last_fetched_timestamp
 ) VALUES (
 ?,?,?,?,?,?,?,?,?,?,?,?,?,?,
-?,?,?,?,?,?,?,?,?,'id',?,?,?,?,?,?,?,?,?,?,?,?,?)`,
+?,?,?,?,?,?,?,?,?,'id',?,?,?,?,?,?,?,?,?,?,?,?)`,
 		s.Address,
 		s.Currency,
 		s.CurrentNetwork,
@@ -190,7 +189,6 @@ INSERT INTO settings (
 		s.AutoRefreshTokensEnabled,
 		// Default the news feed last fetched timestamp to now
 		time.Now().Unix(),
-		s.BackupFetched,
 	)
 	if err != nil {
 		return err
@@ -405,7 +403,7 @@ func (db *Database) GetSettings() (Settings, error) {
 		syncing_on_mobile_network, default_sync_period, use_mailservers, messages_from_contacts_only, usernames, appearance,
 		profile_pictures_show_to, profile_pictures_visibility, wallet_root_address, wallet_set_up_passed, wallet_visible_tokens,
 		waku_bloom_filter_mode, webview_allow_permission_requests, current_user_status, send_status_updates, gif_recents,
-		gif_favorites, opensea_enabled, last_backup, backup_enabled, auto_message_enabled, gif_api_key,
+		gif_favorites, opensea_enabled, auto_message_enabled, gif_api_key,
 		test_networks_enabled, mutual_contact_enabled, profile_migration_needed, wallet_token_preferences_group_by_community, url_unfurling_mode,
 		mnemonic_was_not_shown, wallet_show_community_asset_when_sending_tokens, wallet_display_assets_below_balance,
 		wallet_display_assets_below_balance_threshold, wallet_collectible_preferences_group_by_collection, wallet_collectible_preferences_group_by_community,
@@ -475,8 +473,6 @@ func (db *Database) GetSettings() (Settings, error) {
 		&sqlite.JSONBlob{Data: &s.GifRecents},
 		&sqlite.JSONBlob{Data: &s.GifFavorites},
 		&s.OpenseaEnabled,
-		&s.LastBackup,
-		&s.BackupEnabled,
 		&s.AutoMessageEnabled,
 		&s.GifAPIKey,
 		&s.TestNetworksEnabled,
@@ -625,40 +621,8 @@ func (db *Database) ShouldBroadcastUserStatus() (result bool, err error) {
 	return result, err
 }
 
-func (db *Database) BackupEnabled() (result bool, err error) {
-	err = db.makeSelectRow(BackupEnabled).Scan(&result)
-	if err == sql.ErrNoRows {
-		return true, nil
-	}
-	return result, err
-}
-
 func (db *Database) AutoMessageEnabled() (result bool, err error) {
 	err = db.makeSelectRow(AutoMessageEnabled).Scan(&result)
-	if err == sql.ErrNoRows {
-		return true, nil
-	}
-	return result, err
-}
-
-func (db *Database) LastBackup() (result uint64, err error) {
-	err = db.makeSelectRow(LastBackup).Scan(&result)
-	if err == sql.ErrNoRows {
-		return 0, nil
-	}
-	return result, err
-}
-
-func (db *Database) SetLastBackup(time uint64) error {
-	return db.SaveSettingField(LastBackup, time)
-}
-
-func (db *Database) SetBackupFetched(fetched bool) error {
-	return db.SaveSettingField(BackupFetched, fetched)
-}
-
-func (db *Database) BackupFetched() (result bool, err error) {
-	err = db.makeSelectRow(BackupFetched).Scan(&result)
 	if err == sql.ErrNoRows {
 		return true, nil
 	}
