+++ protocol/communities_messenger_shared_member_address_test.go
@@ -1,6 +1,7 @@
 package protocol
 
 import (
+	"fmt"
 	"testing"
 	"time"
 
@@ -49,7 +50,6 @@ func (s *MessengerCommunitiesSharedMemberAddressSuite) SetupTest() {
 	s.owner = s.newMessenger(ownerPassword, []string{ownerAddress}, "owner", []Option{})
 
 	s.bob = s.newMessenger(bobPassword, []string{bobAddress}, "bob", []Option{})
-	s.bob.EnableBackedupMessagesProcessing()
 
 	s.alice = s.newMessenger(alicePassword, []string{aliceAddress1, aliceAddress2}, "alice", []Option{})
 
@@ -493,8 +493,8 @@ func (s *MessengerCommunitiesSharedMemberAddressSuite) TestResendSharedAddresses
 	backupMessage, err := s.bob.backupCommunity(community, clock)
 	s.Require().NoError(err)
 
-	err = s.bob.HandleBackup(s.bob.buildMessageState(), backupMessage, nil)
-	s.Require().NoError(err)
+	errs := s.bob.handleLocalBackupCommunities(s.bob.buildMessageState(), backupMessage.Communities)
+	s.Require().Len(errs, 0, fmt.Sprintf("expected no errors while handling backup communities. Errors: %v", errs))
 
 	// Owner will receive the request for addresses and send them back to Bob
 	response, err := WaitOnMessengerResponse(
@@ -576,8 +576,8 @@ func (s *MessengerCommunitiesSharedMemberAddressSuite) TestTokenMasterReceivesMe
 	backupMessage, err := s.bob.backupCommunity(community, clock)
 	s.Require().NoError(err)
 
-	err = s.bob.HandleBackup(s.bob.buildMessageState(), backupMessage, nil)
-	s.Require().NoError(err)
+	errs := s.bob.handleLocalBackupCommunities(s.bob.buildMessageState(), backupMessage.Communities)
+	s.Require().Len(errs, 0, fmt.Sprintf("expected no errors while handling backup communities. Errors: %v", errs))
 
 	// Owner will receive the request for addresses and send requests to join with revealed
 	// addresses to token master
