+++ protocol/messenger_backup_handler.go
@@ -2,15 +2,12 @@ package protocol
 
 import (
 	"database/sql"
-	"sync"
-	"time"
 
 	"go.uber.org/zap"
 
 	"github.com/golang/protobuf/proto"
 
 	utils "github.com/status-im/status-go/common"
-	"github.com/status-im/status-go/crypto/types"
 	"github.com/status-im/status-go/images"
 	messagingtypes "github.com/status-im/status-go/messaging/types"
 	"github.com/status-im/status-go/multiaccounts/errors"
@@ -29,102 +26,6 @@ const (
 	SyncWakuSectionKeyKeypairs          = "keypairs"
 	SyncWakuSectionKeyWatchOnlyAccounts = "watchOnlyAccounts"
 )
-
-const backupSyncingNotificationID = "BACKUP_SYNCING"
-
-type FetchingBackedUpDataTracking struct {
-	LoadedItems map[uint32]bool
-	TotalNumber uint32
-}
-
-type BackupFetchingStatus struct {
-	dataProgress           map[string]FetchingBackedUpDataTracking
-	lastKnownMsgClock      uint64
-	fetchingCompleted      bool
-	fetchingCompletedMutex sync.Mutex
-}
-
-func (m *Messenger) HandleBackup(state *ReceivedMessageState, message *protobuf.Backup, statusMessage *messagingtypes.Message) error {
-	if !m.processBackedupMessages {
-		return nil
-	}
-
-	errors := m.handleBackup(state, message)
-	if len(errors) > 0 {
-		for _, err := range errors {
-			m.logger.Warn("failed to handle Backup", zap.Error(err))
-		}
-		return errors[0]
-	}
-	return nil
-}
-
-// TODO remove this function once we do the Waku backup removal
-func (m *Messenger) handleBackup(state *ReceivedMessageState, message *protobuf.Backup) []error {
-	var errors []error
-
-	err := m.handleBackedUpProfile(message.Profile, message.Clock)
-	if err != nil {
-		errors = append(errors, err)
-	}
-
-	for _, contact := range message.Contacts {
-		err = m.HandleSyncInstallationContactV2(state, contact, nil)
-		if err != nil {
-			errors = append(errors, err)
-		}
-	}
-
-	err = m.handleSyncChats(state, message.Chats)
-	if err != nil {
-		errors = append(errors, err)
-	}
-
-	communityErrors := m.handleSyncedCommunities(state, message)
-	if len(communityErrors) > 0 {
-		errors = append(errors, communityErrors...)
-	}
-
-	err = m.HandleBackedUpSettings(message.Setting)
-	if err != nil {
-		errors = append(errors, err)
-	}
-
-	err = m.handleKeypair(message.Keypair)
-	if err != nil {
-		errors = append(errors, err)
-	}
-
-	err = m.HandleWatchOnlyAccount(message.WatchOnlyAccount)
-	if err != nil {
-		errors = append(errors, err)
-	}
-
-	// Send signal about applied backup progress
-	if m.config.messengerSignalsHandler != nil {
-		response := wakusync.WakuBackedUpDataResponse{
-			Clock: message.Clock,
-		}
-
-		response.AddFetchingBackedUpDataDetails(SyncWakuSectionKeyProfile, message.ProfileDetails)
-		response.AddFetchingBackedUpDataDetails(SyncWakuSectionKeyContacts, message.ContactsDetails)
-		response.AddFetchingBackedUpDataDetails(SyncWakuSectionKeyCommunities, message.CommunitiesDetails)
-		response.AddFetchingBackedUpDataDetails(SyncWakuSectionKeySettings, message.SettingsDetails)
-		response.AddFetchingBackedUpDataDetails(SyncWakuSectionKeyKeypairs, message.KeypairDetails)
-		response.AddFetchingBackedUpDataDetails(SyncWakuSectionKeyWatchOnlyAccounts, message.WatchOnlyAccountDetails)
-
-		err = m.updateBackupFetchProgress(message, &response, state)
-		if err != nil {
-			errors = append(errors, err)
-		}
-
-		m.config.messengerSignalsHandler.SendWakuFetchingBackupProgress(&response)
-	}
-
-	state.Response.BackupHandled = true
-
-	return errors
-}
 
 func (m *Messenger) handleLocalBackup(state *ReceivedMessageState, message *protobuf.MessengerLocalBackup) []error {
 	var errors []error
@@ -152,154 +53,6 @@ func (m *Messenger) handleLocalBackup(state *ReceivedMessageState, message *prot
 	}
 
 	return errors
-}
-
-func (m *Messenger) updateBackupFetchProgress(message *protobuf.Backup, response *wakusync.WakuBackedUpDataResponse, state *ReceivedMessageState) error {
-	if m.backedUpFetchingStatus == nil {
-		return nil
-	}
-
-	m.backedUpFetchingStatus.fetchingCompletedMutex.Lock()
-	defer m.backedUpFetchingStatus.fetchingCompletedMutex.Unlock()
-
-	if m.backedUpFetchingStatus.fetchingCompleted {
-		return nil
-	}
-
-	if m.backedUpFetchingStatus.lastKnownMsgClock > message.Clock {
-		return nil
-	}
-
-	if m.backedUpFetchingStatus.lastKnownMsgClock < message.Clock {
-		// Reset the progress tracker because we have access to a more recent copy of the backup
-		m.backedUpFetchingStatus.lastKnownMsgClock = message.Clock
-		m.backedUpFetchingStatus.dataProgress = make(map[string]FetchingBackedUpDataTracking)
-		for backupName, details := range response.FetchingBackedUpDataDetails() {
-			m.backedUpFetchingStatus.dataProgress[backupName] = FetchingBackedUpDataTracking{
-				LoadedItems: make(map[uint32]bool),
-				TotalNumber: details.TotalNumber,
-			}
-		}
-
-		if len(m.backedUpFetchingStatus.dataProgress) == 0 {
-			return nil
-		}
-	}
-
-	// Evaluate the progress of the backup
-
-	// Set the new items before evaluating
-	for backupName, details := range response.FetchingBackedUpDataDetails() {
-		m.backedUpFetchingStatus.dataProgress[backupName].LoadedItems[details.DataNumber] = true
-	}
-
-	for _, tracker := range m.backedUpFetchingStatus.dataProgress {
-		if len(tracker.LoadedItems)-1 < int(tracker.TotalNumber) {
-			// have not received everything yet
-			return nil
-		}
-	}
-
-	m.backedUpFetchingStatus.fetchingCompleted = true
-
-	// Update the AC notification and add it to the response
-	notification, err := m.persistence.GetActivityCenterNotificationByID(types.FromHex(backupSyncingNotificationID))
-	if err != nil {
-		return err
-	}
-
-	if notification == nil {
-		return nil
-	}
-
-	notification.UpdatedAt = m.GetCurrentTimeInMillis()
-	notification.Type = ActivityCenterNotificationTypeBackupSyncingSuccess
-	_, err = m.persistence.SaveActivityCenterNotification(notification, true)
-	if err != nil {
-		return err
-	}
-
-	state.Response.AddActivityCenterNotification(notification)
-	return nil
-}
-
-func (m *Messenger) startBackupFetchingTracking(response *MessengerResponse) error {
-	// Add an acivity center notification to show that we are fetching back up messages
-	notification := &ActivityCenterNotification{
-		ID:        types.FromHex(backupSyncingNotificationID),
-		Type:      ActivityCenterNotificationTypeBackupSyncingFetching,
-		Timestamp: m.getTimesource().GetCurrentTime(),
-		Read:      false,
-		Deleted:   false,
-		UpdatedAt: m.GetCurrentTimeInMillis(),
-	}
-	err := m.addActivityCenterNotification(response, notification, nil)
-
-	if err != nil {
-		return err
-	}
-
-	// Add a timeout to mark the backup syncing as failed after 1 minute and 30 seconds
-	m.shutdownWaitGroup.Add(1)
-	go func() {
-		defer utils.LogOnPanic()
-		defer m.shutdownWaitGroup.Done()
-		m.watchBackupFetching()
-	}()
-
-	return nil
-}
-
-func (m *Messenger) watchBackupFetching() {
-	select {
-	case <-m.ctx.Done():
-		return
-	case <-time.After(90 * time.Second):
-		m.backupFetchingTimeout()
-	}
-}
-
-func (m *Messenger) backupFetchingTimeout() {
-	if m.backedUpFetchingStatus == nil {
-		return
-	}
-
-	m.backedUpFetchingStatus.fetchingCompletedMutex.Lock()
-	defer m.backedUpFetchingStatus.fetchingCompletedMutex.Unlock()
-
-	if m.backedUpFetchingStatus.fetchingCompleted {
-		// Nothing to do, the fetching has completed successfully
-		return
-	}
-	// Update the AC notification to the failure state
-	notification, err := m.persistence.GetActivityCenterNotificationByID(types.FromHex(backupSyncingNotificationID))
-	if err != nil {
-		m.logger.Error("failed to get activity center notification", zap.Error(err))
-		return
-	}
-	if notification == nil {
-		return
-	}
-
-	notification.UpdatedAt = m.GetCurrentTimeInMillis()
-	if len(m.backedUpFetchingStatus.dataProgress) == 0 {
-		notification.Type = ActivityCenterNotificationTypeBackupSyncingFailure
-	} else {
-		notification.Type = ActivityCenterNotificationTypeBackupSyncingPartialFailure
-	}
-	_, err = m.persistence.SaveActivityCenterNotification(notification, true)
-	if err != nil {
-		m.logger.Error("failed to save activity center notification", zap.Error(err))
-		return
-	}
-
-	if m.config.messengerSignalsHandler == nil {
-		return
-	}
-
-	resp := &MessengerResponse{}
-	resp.AddActivityCenterNotification(notification)
-	m.config.messengerSignalsHandler.MessengerResponse(resp)
 }
 
 func (m *Messenger) handleBackedUpProfile(message *protobuf.BackedUpProfile, backupTime uint64) error {
@@ -411,6 +164,7 @@ func (m *Messenger) handleBackedUpProfile(message *protobuf.BackedUpProfile, bac
 	return err
 }
 
+// TODO move this to the AccountsService
 func (m *Messenger) HandleBackedUpSettings(message *protobuf.SyncSetting) error {
 	if message == nil {
 		return nil
@@ -457,64 +211,6 @@ func (m *Messenger) HandleBackedUpSettings(message *protobuf.SyncSetting) error 
 	return nil
 }
 
-func (m *Messenger) handleKeypair(message *protobuf.SyncKeypair) error {
-	if message == nil {
-		return nil
-	}
-
-	multiAcc, err := m.multiAccounts.GetAccount(message.KeyUid)
-	if err != nil {
-		return err
-	}
-	// If user is recovering his account via seed phrase, but the backed up messages indicate that the profile keypair
-	// is a keycard related profile, then we need to remove related profile keycards (only profile, other keycards should remain).
-	if multiAcc != nil && multiAcc.KeyUID == message.KeyUid && !multiAcc.RefersToKeycard() && len(message.Keycards) > 0 {
-		message.Keycards = []*protobuf.SyncKeycard{}
-	}
-
-	keypair, err := m.handleSyncKeypair(message, false, nil)
-	if err != nil {
-		if err == ErrTryingToStoreOldKeypair {
-			return nil
-		}
-		return err
-	}
-
-	if m.config.messengerSignalsHandler != nil {
-		kpResponse := wakusync.WakuBackedUpDataResponse{
-			Keypair: keypair.CopyKeypair(),
-		}
-
-		m.config.messengerSignalsHandler.SendWakuBackedUpKeypair(&kpResponse)
-	}
-
-	return nil
-}
-
-func (m *Messenger) HandleWatchOnlyAccount(message *protobuf.SyncAccount) error {
-	if message == nil {
-		return nil
-	}
-
-	acc, err := m.handleSyncWatchOnlyAccount(message, true)
-	if err != nil {
-		if err == ErrTryingToStoreOldWalletAccount {
-			return nil
-		}
-		return err
-	}
-
-	if m.config.messengerSignalsHandler != nil {
-		response := wakusync.WakuBackedUpDataResponse{
-			WatchOnlyAccount: acc,
-		}
-
-		m.config.messengerSignalsHandler.SendWakuBackedUpWatchOnlyAccount(&response)
-	}
-
-	return nil
-}
-
 func syncInstallationCommunitiesSet(communities []*protobuf.SyncInstallationCommunity) map[string]*protobuf.SyncInstallationCommunity {
 	ret := map[string]*protobuf.SyncInstallationCommunity{}
 	for _, c := range communities {
@@ -525,24 +221,6 @@ func syncInstallationCommunitiesSet(communities []*protobuf.SyncInstallationComm
 		}
 	}
 	return ret
-}
-
-// TODO remove this function once we do the Waku backup removal
-func (m *Messenger) handleSyncedCommunities(state *ReceivedMessageState, message *protobuf.Backup) []error {
-	var errors []error
-	for _, syncCommunity := range syncInstallationCommunitiesSet(message.Communities) {
-		err := m.handleSyncInstallationCommunity(state, syncCommunity)
-		if err != nil {
-			errors = append(errors, err)
-		}
-
-		err = m.requestCommunityKeysAndSharedAddresses(state, syncCommunity)
-		if err != nil {
-			errors = append(errors, err)
-		}
-	}
-
-	return errors
 }
 
 func (m *Messenger) handleLocalBackupCommunities(state *ReceivedMessageState, communities []*protobuf.SyncInstallationCommunity) []error {
