+++ protocol/messenger_testing_utils.go
@@ -64,7 +64,6 @@ func (m *MessengerSignalsHandlerMock) SendWakuBackedUpKeypair(*wakusync.WakuBack
 func (m *MessengerSignalsHandlerMock) SendWakuBackedUpWatchOnlyAccount(*wakusync.WakuBackedUpDataResponse) {
 }
 
-func (m *MessengerSignalsHandlerMock) BackupPerformed(uint64)                    {}
 func (m *MessengerSignalsHandlerMock) HistoryArchivesProtocolEnabled()           {}
 func (m *MessengerSignalsHandlerMock) HistoryArchivesProtocolDisabled()          {}
 func (m *MessengerSignalsHandlerMock) CreatingHistoryArchives(string)            {}
