+++ protocol/protobuf/application_metadata_message.proto
@@ -50,7 +50,7 @@ message ApplicationMetadataMessage {
     ANONYMOUS_METRIC_BATCH = 33;
     SYNC_CHAT_REMOVED = 34;
     SYNC_CHAT_MESSAGES_READ = 35;
-    BACKUP = 36;
+    BACKUP = 36 [deprecated=true];
     SYNC_ACTIVITY_CENTER_READ = 37;
     SYNC_ACTIVITY_CENTER_ACCEPTED = 38;
     SYNC_ACTIVITY_CENTER_DISMISSED = 39;
