+++ services/ext/service.go
@@ -207,9 +207,6 @@ func (s *Service) InitProtocol(params InitProtocolParams) error {
 		return err
 	}
 	s.messenger = messenger
-	if s.config.ProcessBackedupMessages {
-		s.messenger.EnableBackedupMessagesProcessing()
-	}
 
 	// Be mindful of adding more initialization code, as it can easily
 	// impact login times for mobile users. For example, we avoid calling
